s32g: wkpu: Correct "reg" property interpretation
Use fdt_get_reg_props_by_index to read "reg" property. It will take
into consideration the number of cells used for address and size
specification.

Issue: ALB-4294, ALB-5510

+++ plat/nxp/s32g/include/s32g_dt.h
@@ -7,11 +7,13 @@
 #ifndef S32G_DT_H
 #define S32G_DT_H
 
+#include <stdint.h>
+
 #define DT_DISABLED	0
 #define DT_ENABLED	1
 
 struct dt_node_info {
-	uint32_t base;
+	uintptr_t base;
 	int32_t clock;
 	int32_t reset;
 	uint32_t status;
